+++ changelog.md
@@ -11,6 +11,7 @@
 - [#4880](https://github.com/ignite/cli/pull/4880) Add possibility to scaffold underscore package.
 - [#4878](https://github.com/ignite/cli/pull/4878) Improve the `xast` package readability.
 - [#4883](https://github.com/ignite/cli/pull/4883) Remove ibc keys placeholders.
+- [#4884](https://github.com/ignite/cli/pull/4884) Remove autocli placeholders.
 
 ### Fixes