+++ changelog.md
@@ -12,6 +12,7 @@
 - [#4878](https://github.com/ignite/cli/pull/4878) Improve the `xast` package readability.
 - [#4883](https://github.com/ignite/cli/pull/4883) Remove ibc keys placeholders.
 - [#4884](https://github.com/ignite/cli/pull/4884) Remove autocli placeholders.
+- [#1616](https://github.com/ignite/cli/issues/1616) Prevent having pointer for custom type.
 
 ### Fixes