+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [`v29.9.0`](https://github.com/ignite/cli/releases/tag/v29.9.0)
+
 ## Features
 
 - [#4877](https://github.com/ignite/cli/pull/4877) Remove app config and ibc add route placeholders.